Add sentinel at index 0 to SN list to simplify logic
Currently the service node list in the smart contract is managed by
a circular linked list where LIST_END (U64_MAX) is reserved as the head
of the linked list. We also reserve slot 0 (by setting nextServiceNodeID to 1)
and make it unallocatable.

We can avoid wasting a slot by placing a sentinel at slot 0. The logic
plays out similarly in that the sentinel points to itself as if it were
circular.

In addition to that, since it's a doubly linked-list, there are common patterns
used to assign and delete from the list that I've opted for. These patterns
simplify the code in our instance, `seedPublicKeyList` was declaring various
temporary variables to track list state which were not necessary.

I've wrapped the append and delete from list into functions with the equivalent
pattern as you would see it in C++ code, e.g.

```
func serviceNodeAdd(...) {
    // The following is the insertion pattern in a doubly-linked list
    // with sentinel at index 0
    //
    // ```c
    // node->next       = sentinel;
    // node->prev       = sentinel->prev;
    // node->next->prev = node;
    // node->prev->next = node;
    // ```
    ...
}
```

I've added unit tests to verify the state of the service node list and
its pointed-ness matches what we manually derived in C++'s ServiceNodeList
with the new service node functionality exposed in C++.