Fix onlyOperator w/tx.origin allowing auth to be bypassed
You can maliciously construct a tx.origin that is the operator by
getting them to interact with a malicious contract which then forwards
the request to the target contract.

In our instance, this allows someone else to finalize or cancel the node
on behalf of the operator which is a mild inconvenience but one that we
can block by checking `msg.sender` instead of `tx.origin`.

We note that `tx.origin` is necessary in the constructor because the
intended workflow is to create a contract through the contribution
factory (which allows applications to enumerate service nodes accepting
contribution by parsing the logs of that contract) in which case the
original address of the TX that interacted with the factory is
identified as the service node operator.