Fix reward pool not updating lastPaidOutTime & double paying
`payoutRelease` was not updating the lastPaidOutTime which meant that it was
paying from the genesis of the contract creation everytime
`payoutRelease` was requested.

Additionally, `calculateRelease` is the amount of tokens released in
totality up to the specified timestamp. This value was incorrectly being
summed onto the `totalPaidOut` which also would mean paying out all the
rewards from the genesis of the contract creation.

I fixed this accordingly and adjusted the unit test to expose this bug.