Use contributeAddresses array length to track numContributors
We can defer to the actual data structure storing the contributors to
determine the number of contributors. This data is already stored in the
contract in a storage slot that we can reuse. It also avoids having to
synchronise state and the potential for them to go out of sync.