Use safeTransfer or check transfer result via slither
In the places that slither identified that we should validate the
transfer result or use `safeTransfer` there were no real identifiable
problems because there was no state changing in the smart contract
except for the token balance itself.

We also lock down the ERC20 token we interact with to the SENT token
address which will be initialised with OpenZeppelin's implementation so
encountering malicious code in the transfer is unlikely that might cause
it to fail but allow contract side effects to continue to be updated.

However it doesn't cost much to defer to a `safeTransfer` call which
essentially wraps the `transfer` with a `require` for us so we do so to
fix the warning.