Zellic 3.5: Wrong input length to static call
The add function of BN256G1 is calling the precompiled contract
performing an elliptic curve point addition. It uses a staticcall opcode
directly in assembly. The opcode expects an input length as a parameter.
In this function, it is set to 0xc0 but the input array is only 4 * 32
bytes, so it should be 0x80 instead.

The mul function has a similar mismatch with the input length set to
0x80, whereas there are only three elements in the input array.