Zellic 4.3: Misleading functions' name
Functions named with suffix Jacobian in BN254G2 library are in fact not
working with Jacobian coordinates but with projective coordinates. The
point-addition algorithm for G2 implemented by the function
_ECTwistAddJacobian is described by formula (3) of this paper,
"Efficient Elliptic Curve Exponentiation Using Mixed Coordinates".

The point-doubling algorithm implemented by function
_ECTwistDoubleJacobian is the implementation of formula (4) of the same
paper. The point-doubling algorithm was later optimized in
[hyperelliptic.org](https://hyperelliptic.org/EFD/g1p/auto-shortw-projective.html#addition-add-1998-cmo-2).
The library may benefit from those improvements. In the same way, the
function _fromJacobian transforms projective into affine coordinates.

The function _ECTwistMulByCofactorJacobian is an implementation of the
algorithm from the paper "Faster Hashing to G2" and is also not
working on Jacobian coordinates. The naming is misleading, and the lack
of reference to the original algorithms may be error-prone for later
developments