Zellic 4.2: Gas Savings
In the FQ2Sqrt function of BN256G2, a square root computation in the
base field Fp can be saved. If the number has no imaginary part (i.e.,
x2 == 0) and the number is not a square, then the second square root
computation _sqrt(FIELD_MODULUS - x1) is not necessary as it will return
the same value t1 as for the previous square root computation.