Zellic 3.2: Lack of proof-of-possession verification
Before aggregating a BLS public key, a proof of possession is required
by the contract. It consists in a signature of the domain separation
tag, the public key together with the caller address, and the
service-node public key. This allows proving to the contract that the
service node knows the correct private key matching the public key
submitted. It means that a node knows the private key xi of its public
key Qi such that Qi = xi · G1, where G1 is the generator of the group
G1.

However, the function seedPublicKeyList does not use a proof of
possession before adding a list of public keys to the aggregate key:

```
/// @notice Seeds the public key list with an initial set of service nodes.
///
/// @dev This should be called before the hardfork by the foundation to
/// ensure the public key list is ready to operate.
///
/// @param pkX Array of X-coordinates for the public keys.
/// @param pkY Array of Y-coordinates for the public keys.
/// @param amounts Array of amounts that the service node has staked,
/// associated with each public key.
function seedPublicKeyList(
    uint256[] calldata pkX,
    uint256[] calldata pkY,
    uint256[] calldata amounts
) external onlyOwner {
    if (pkX.length != pkY.length || pkX.length != amounts.length) {
        revert ArrayLengthMismatch();
    }
    for (uint256 i = 0; i < pkX.length; i++) {
        BN256G1.G1Point memory pubkey = BN256G1.G1Point(pkX[i], pkY[i]);
        uint64 allocID = serviceNodeAdd(pubkey);
        _serviceNodes[allocID].deposit = amounts[i];
        emit NewSeededServiceNode(allocID, pubkey);
    }
    updateBLSNonSignerThreshold();
}
```

Therefore, a rogue key attack is possible. For an aggregate key Q = ∑i
Qi, the owner of the contract can add a new rogue key Qr = xr · G − ∑i Qi to the
aggregate key. Then the new aggregate key verifies signatures signed
only by their key xr but not the other previous keys. This means that
messages signed only by their key will be seen as valid even if they
were not signed by the other public keys

---

This vulnerability is acknowledged and will be handled off-band via our
C++ code that will be responsible for generating the parameters for the
migration into Arbitrum. As part of the migration, cryptographic proofs
will have to generated by the pre-existing network of Session Nodes
containing a proof-of-possession for the secret component of the BLS
public key the node declares.

Additionally after the contract is started by invoking `start` this
function is disabled to prevent rogue keys being added by the owner
after seeding. The seeding of the list may take multiple transactions
due to gas restrictions such that we disable seeding after we start as
opposed to directly after calling the seeding function..