Zellic 3.1: Potential vulnerability in _aggregatePubkey update mechanism
Upon examining the code where the _aggregatePubkey is updated through
the addBLSPublicKey function, we have identified a potentially
significant scenario that requires attention. Assuming that the DEX swap
pair that includes SENT tokens has a large liquidity in the future,
there exists a potential vulnerability.

Specifically, if a flash loan is used to acquire enough tokens to
constitute more than a 2/3 majority, and these tokens are then staked,
the _aggregatePubKey could be updated immediately.

The code involved is shown below.

```
function serviceNodeAdd(BN256G1.G1Point memory pubkey) internal returns (uint64 result) {
    [...]
    if (totalNodes == 1) {
    _aggregatePubkey = pubkey;
    } else {
    _aggregatePubkey = BN256G1.add(_aggregatePubkey, pubkey);
    }
    return result;
}
```

This scenario could enable arbitrary reward updates using the
updateRewardsBalance function.

---

The premise of this attack requires several operating conditions:

  1. When the network has more than 300 honest nodes (nb; there are
  currently ~2200 nodes active), no signature produced by the attacker
  will be permitted due to the BLS non-signing threshold. That is, if
  there are more than `blsNonSignerThreshold` nodes- which at the limit
  of the network is 300 nodes did not participate in the signing of the
  signature the smart contract will reject the request as codified by the
  modifier `hasEnoughSigners`

  2. There must be sufficient liquidity on the active market to fund the
  flash-loan as acknowledge by Zellic.

  3. When the network has between 10-300 nodes the attacker would need
  to acquire 20-600 pubkeys which would fail to be added due to gas limits.

  4. The presence of a proof-of-possession on node registration currently costs
  around 1.5m gas. Adding 20 pubkeys would put you close or over the gas limits
  of the block.

There are various quirks with the feasibility of this exploit but we
acknowledge that it's sensible to have restrictions on the upper limit
of aggregate pubkey updates there are in a single block and potentially
restrictions between modifying and updating the aggregate pubkey in the
same transaction.

At the bounds of the minimum node limit to be susceptible to attack, if
the network has less than 10 nodes the attacker could feasibly register
20 nodes to pass the threshold for a signature to pass. In this
situation it is agreed upon that there are potential vulnerabilities in
a network with a weakened topology. In this scenario the security model
of Session would be compromised and such situation would need to be assessed
with a fall-back to manual contract operations.

In our solution we have added a rate limit that allows up to
`max(5, 2% of the network)` pubkeys to be aggregated into the contract's
`_aggregatePubkey` which mitigates the exploit of overtaking the network
within the same block using a flash loan.