Zellic 3.4: Bias in hashToField function
The function hashToField is used to hash a uint256 value toFp. The
resulting value is later mapped to a point on an elliptic curve with the
function mapToG2. The message is first hashed with the keccak256
function and then transformed to a value in the finite field with the
function maskBits:

```solidity
// This matches mcl maskN, this only takes the 254 bits for the field, if it is
still greater than the field then take the 253 bits
function maskBits(uint256 input) internal pure returns (uint256) {
    uint256 mask = ~uint256(0) - 0xC0;
    if (byteSwap(input & mask) >= FIELD_MODULUS) {
        mask = ~uint256(0) - 0xE0;
    }
    return input & mask;
}
```

The two first bits of the value are set to zero, and then if the value
is still bigger than p, the next bit is masked to zero. It means that
the values between p and 2^254−1 are mapped to a value between p−2^253
and 2^253−1. Values between zero and p are left unchanged, resulting in
values in the range [p−2^253, 2^253−1] having twice the probability to
be chosen as output.

---

Our solution to this is to implement RFC9380. The implementation has
largely been taken from Hopr's crypto implementation of RFC9380 tweaked
for the BLS usecase.