Remove timestamp argument from RewardRatePool query functions
Requiring the timestamp here causes problems with oxend because it means
we can't realistically pre-fetch rewards amounts until we get the
timestamp from oxen.

That's a problem because we can't pre-fetch the reward data for
verification (because it needs the timestamp that goes into the block),
which means in order to validate blocks we'd have to stop during block
processing and make (and wait for) a contract call to come back.

Removing the timestamp makes this a lot easier because it lets us query
the reward rate in advance by only querying every few minutes (at
predefined height intervals the network agrees on) and because it now
depends on the block timestamp, the value will be the same for anyone
fetching it.