Avoid serviceNodesLength() in favour of precomputed totalNodes
serviceNodeLengths() is relatively expensive in that it has to walk the
entire linked list (which involves a lot of storage accesses), and so we
shouldn't be using it in general when the much cheaper totalNodes is
available to us.

It isn't entirely clear that the serviceNodesLength and
updateServiceNodesLength have a useful purpose (totalNodes should always
be maintained at the correct value anyway), but they are perhaps useful
backup functionality for a serious contract mishap, and leaving them
doesn't *hurt* anything.