Add admin function for stagenet temporarily
This function will be removed immediately after upgrading the
testnet contract because the original mainnet version was deployed.
Trying to deploy `test/TestnetServiceNodeRewards.sol` doesn't work
because the hash the contract would have been deployed at is not present
on stagenet (I believe this address is the proxy contract address
which means that the testnet version of the contract was _not_ deployed).

So, what I've opted for is to migrate the admin function to the main
contract, deploy it, and then remove the admin contract and give a big
disclaimer on what this function is doing here.

This will be committed to the repository to provide a perma-link for
documentation purposes of the list of upgrades that have been deployed on
stagenet and the commit will be reverted there-after.