Inject missing node from stagenet mishap
In stagenet we used the admin function to delete a bunch of nodes. In
the snapshotting process a node was deleted that was not meant to be
deleted causing it to persist in the stagenet blockchain but not the
smart contract. This commit addresses that by injecting the node back
into the list.