Cosmetic fixes to confusing event parameter names
A few events have an address parameter named "recipient" which was
confusing as the events in question have little to do with receiving
anything.  This updates them to a more descriptive name.

This was spreading into oxen-core, where attempting to preserve naming
consistency with the smart contract led to similar confusion of the
values being captured in the eth transactions.

(This is purely cosmetic as no types in the events are being changed and
event signatures depend only on parameter types, not parameter names).