Exponential Moving Average Check
Adds a check to the claim rewards function to rate limit the claim
rewards function. This makes an upper limit to the number of tokens that
can be claimed in a day. Which in the event of an extremely high number
of tokens being withdrawn will allow the owner time to pause the
contract and investigate.

The exponential moving average uses a previous value, a time since
last claim, a half life and a threshold.

The claimEMA is an adjusted amount summarising the history of
previous claims. When a new claim is made the previous value is adjusted
by an alpha value which modifys the previous value based on the length
of time since the last claim

`alpha = 1 - (1/2) ^ (timeDiff / halfLife)`

This alpha will between 1 and zero and the graph takes on the form of
`sqrt(x)` with alpha approaching 1 as `timeDiff` increases over `halfLife`.

Taking this alpha it is then applied using the following formula:

`claimEMA = newClaimAmount + oldEMA * (1 - alpha)`

This means that the claimEMA will only partially use the old value and
less so if it has been a long time since the last claim.

We then compare this claimEMA to our threshold amount to make sure that
users havent claimed too many tokens in too short of a period.