Don't impose new-contributor requirements on top-ups
This removes the requirement that new contributions from existing
contributors satisfy the minimum staking requirement.  Such contributors
will already satisfy the minimum requirement (because it would have been
checked by the initial contribution), and so this doesn't need to be
checked.  *Not* checking it allows contributors to increase their stake
in a pending SN with any amount they like.

The current contract behaviour was mimicking the Oxen behaviour which
*does* impose the minimum on top-ups, but that was because of a
implementation detail (realistically, a limitation, perhaps
categorizable as a bug) in Oxen where the minimum was defined not by the
number of distinct contributors, but rather than the number of locked
key images which always increased with each top-up, and so Oxen top-ups
really did take up a contribution spot.