Add more setters for recipient/liquidation ratio
Change the revert error for the ratio, rewards is emitted for the error in claim
rewards. Reusing it in the constructor is potentially confusing that
it's not the rewards claim failed but initialisation. We have a catch
all error, PositiveNumberRequirement that we can use here to indicate
such errors and disambiguate it from an _actual_ rewards claim error.