Remove getters that are auto-generated from SN rewards
Public variables have auto-generated getters defined for them. Public
variables can't be modified by default, they are read-only, which means
that the private+getter code we have was redundant.

Getters for structs have remained because solidity auto-generates
a getter that desugars the field into a tuple which is less useful than
if we'd manually returned the struct definition itself.