Inline addBLSPublicKey w/ explicit msg.sender
Before we have `addBLSPublicKey` and `_addBLSPublicKey` where
`addBLSPublicKey` was a wrapper over `_addBLSPublicKey` but it snapped
a copy of `msg.sender` to pass into `addBLSPublicKey`.

This is not needed as there's no other function or auxiliary contract
calling `_addBLSPublicKey` such that they are allowed or need to
parameterise the caller (which is the owner) of the node that they're
adding a BLS public key for.

Hence, the wrapper function just adds a layer of indirection for no real
gain so we can inline it and remove that indirection.

On the other hand, `initiateRemoveBLSPublicKey` has 2 functions one that
parameterises the caller and one that passes in `msg.sender`. We have
a testing contract, `TestnetServiceNodeRewards` with an admin function
that wants to used the parameterised caller to force-eject a node from the SNL.

In this case you want to have the 2 APIs to allow the testnet contract
to inherit and invoke directly. This is not the case for
`addBLSPublicKey` so we can reduce the footprint of the contract by
removing the wrapper.