Make finalization configurable, defaults to automatic
Introduce manualFinalize which configures the automatic finalisation
behaviour of the contract.

When ==false (this is the default behaviour of a spawned contract), the
contract will automatically finalize the contract (e.g. register the
node) upon receipt of the last contribution that allows the contract to
meet the staking requirement (even if the last contribution was
conducted by a public contributor!).

When ==true operator has to go and invoke finalize when the staking
requirement is met. Operators can monitor for the `Filled` event that is
emitted when the staking requirement is met to be notified that they
need to finalize their contract.