Don't set beneficiary in reserved to avoid malicious operators
When setting reserved contributors there's no point in setting the
beneficiary there as we only use the staker's address to reserve the
addresses that are initially allowed to contribute to the node.

Removing this also prevents the operator from maliciously manipulating
the beneficiary address (and if the contributor doesn't notice the
operator changed it) and siphon rewards away from them.

Reserved and public contributors can initially set their beneficiary
when they contribute. This means that the only person that is authorised
to change the beneficiary is the wallet contributing and hence removes
risk of malicious operators.

Additionally fix the unit tests to accomodate this change and other breakages.