Zellic 3.3: Some functions can be implemented more efficiently
Below we provide our recommendations for gas optimizations.

1) In several functions, the length of an array can be cached to save
gas in loops. The list of functions where this optimization can be
applied is given below:

In the ServiceNodeContribution contract
    • reserved.length in the _updateReservedContributors function

In the ServiceNodeRewards contract
    • contributors.length in the addBLSPublicKey function
    • _serviceNodes[serviceNodeID].contributors.length in the _initiateRemoveBLSPublicKey function
    • nodes.length and node.contributors.length in the seedPublicKeyList function

Also, the index incrementing for the loop can be made unchecked to
optimize gas usage.

2) The _contributeFunds function can be updated to avoid calling
_updateBeneficiary when the staker is being added for the first time, as
shown in the example below:

```
if (contributions[caller] == 0)
    _contributorAddresses.push(IServiceNodeRewards.Staker(caller, beneficiary));
} else {
    _updateBeneficiary(caller, beneficiary);
}
```

3) The _reset() function can be optimized. When status ==
Status.Finalized, it is unnecessary to call removeAndRefundContributor
for each address in the _contributorAddresses array, since in this
case, the entire array should simply be deleted and the contributions
and contributionTimestamp should be set to zero. If the state is not
Finalized, there is also no need to call removeAndRefundContributor for
every element in the _contributorAddresses just to delete a single
element. Currently, the removeAndRefundContributor function searches for
the staker's address in the _contributorAddresses array, moves elements,
removes the last one, and then transfers funds. In case of status ==
Status.Finalized, we suggest iterating through the _contributorAddresses
array to refund each staker, set the contributions[staker] and
contributionTimestamp[staker] to zero, and finally delete the
_contributorAddresses array. Finally, call the function
_updateReservedContributors with zero contributors.

--

Our proposed fix is as per Zellic's recommendations

1) We cache the array length and used unchecked index increment on all
our contracts (where it is safe to do so, i.e. a predictable known
non-overflowing length).

2) Avoid calling updateBeneficiary on initial contribution to contract

3) Directly handle reset rather than defer to removeAndRefundContributor

We can directly handle a reset by processing the request directly rather
than deferring to removeAndRefundContributor which presupposes that the
intention of the call is to remove exactly 1 contributor which has
side-effects that waste gas as we're going to ultimately delete the
contributor which nullifies said busy work.