Zellic 3.1: Inconsistent status after withdrawing a contribution
In the _contributeFunds function, when the staking requirements are met
and manualFinalize is not set, the status is set to Status.WaitForFinalized,
awaiting the operator to call the finalize function.

```
function _contributeFunds(address caller, address beneficiary, uint256 amount) private {
    [...]
    if (currTotalContribution == stakingRequirement) {
        emit Filled(_serviceNodeParams.serviceNodePubkey, operator);
        status = Status.WaitForFinalized;
    }

    // NOTE: Transfer funds from sender to contract
    emit NewContribution(caller, amount);
    SENT.safeTransferFrom(caller, address(this), amount);

    // NOTE: Auto finalize the node if valid
    if (status == Status.WaitForFinalized && !manualFinalize) {
        _finalize();
    }
}
```
However, during this waiting period, a user can still remove their
contributions, as the withdrawContribution function has no restrictions
during this stage. This leads to a situation where the contract remains
in the Status.WaitForFinalized status, even though the staking
requirement is no longer met after the withdrawal.

```
function withdrawContribution() external {
    if (msg.sender == operator) {
        _reset();
        return;
    }
    uint256 timeSinceLastContribution = block.timestamp - contributionTimestamp[msg.sender];
    if (timeSinceLastContribution < WITHDRAWAL_DELAY)
        revert WithdrawTooEarly(contributionTimestamp[msg.sender], block.timestamp, WITHDRAWAL_DELAY);

    uint256 refundAmount = removeAndRefundContributor(msg.sender);
    if (refundAmount > 0)
        emit WithdrawContribution(msg.sender, refundAmount);
}
```

Impact

The finalize function remains available for calling, even if the staking
requirements are not met. This allows the operator to attempt using the
current collected contributions to add a new service node. However, even
if the contract balance is sufficient, the addBLSPublicKey of the
ServiceNodeRewards contract, which is called during node registration,
will revert. This is because the addBLSPublicKey function performs
a check to ensure that the total contributions from the current
contributors match the stakingRequirement.

```
function addBLSPublicKey(
    BN256G1.G1Point memory blsPubkey,
    BLSSignatureParams memory blsSignature,
    ServiceNodeParams memory serviceNodeParams,
    Contributor[] memory contributors
) external whenNotPaused whenStarted {
    [...]
    for (uint256 i = 0; i < contributors.length; i++)
        totalAmount += contributors[i].stakedAmount;
    if (totalAmount != stakingRequirement)
        revert ContributionTotalMismatch(stakingRequirement, totalAmount);
    [...]
}
```

As the status is still Status.WaitForFinalized, new contributors cannot
add funds via the contributeFunds call, because this function reverts if
the current status is not Status.WaitForOperatorContrib or
Status.OpenForPublicContrib. Thus, the only way to fix the status would
be to reset all the current contributions and request all the
contributors to contribute again. A malicious contributor could thus
carry out the attack numerous times, leading to bad user experience for
the other contributors and the contribution not ever being staked.

Recommendations

We recommend changing the status back to Status.OpenForPublicContrib in
the withdrawContribution call if the current status is
Status.WaitForFinalized and the withdraw amount decreases the
contribution from stakingRequirement.

--

Our proposed fix is as per Zellic's recommendation to revert status back
to being open for contribution on withdraw.