Zellic 3.2: Misleading beneficiary update after finalization
The updateBeneficiary function allows a contributor to change the
current beneficiary address that receives the reward. This function
updates the Staker object in the _contributorAddresses mapping, which is
associated with the contributor.

```
function updateBeneficiary(address newBeneficiary) external {
    _updateBeneficiary(msg.sender, newBeneficiary);
}

function _updateBeneficiary(address stakerAddr, address newBeneficiary) private {
    address desiredBeneficiary = newBeneficiary == address(0) ? stakerAddr :
    newBeneficiary;
    address oldBeneficiary = address(0);
    bool updated = false;
    uint256 length = _contributorAddresses.length;
    for (uint256 i = 0; i < length; i++) {
        IServiceNodeRewards.Staker storage staker = _contributorAddresses[i];
        [...]
        oldBeneficiary = staker.beneficiary;
        staker.beneficiary = desiredBeneficiary;
        break;
    }
    [...]
}
```

Impact

However, the issue arises if the updateBeneficiary function is called
during Status.Finalized, because all information about contributors,
including the beneficiary addresses, has already been provided to the
ServiceNodeRewards contract for new node registration. Any beneficiary
address updated after finalization will not be applied to the node, as
the ServiceNodeRewards contract does not permit updates to the
beneficiary information after registration.

Recommendations

We recommend restricting the updateBeneficiary function during the
Status.Finalized phase to avoid misleading stakers that the beneficiary
addresses has been updated.

--

Our proposed fix is as per the recommendations, we allow a beneficiary
update only when contract is in a valid state.

Updating the beneficiary after the contract is finalized has no effect
because the beneficiaries have already been committed to the contract
and are locked in. This is preventable by checking the status of the
contract before updating the contract.

Note that when the operator contributes, the contract transitions into
`OpenForPublicContrib` immediately before the beneficiary is updated.
This then means that the valid states that are observed for
a beneficiary to be updatable is exactly 2 states,
`OpenForPublicContrib` and `WaitForFinalized`.