Allow beneficiary update only when contract is in valid state
Updating the beneficiary after the contract is finalized has no effect
because the beneficiaries have already been committed to the contract
and are locked in. This is preventable by checking the status of the
contract before updating the contract.

Note that when the operator contributes, the contract transitions into
`OpenForPublicContrib` immediately before the beneficiary is updated.
This then means that the valid states that are observed for
a beneficiary to be updatable is exactly 2 states,
`OpenForPublicContrib` and `WaitForFinalized`.