Allow leave requests events to be repeated
If a leave request gets missed (because of an exceptional circumstance
on the Oxen chain) then the Oxen chain never gets the confirmed leave
request, and so the node cannot complete a normal, graceful exit because
the Oxen SN network will never sign a leave request.  While there *is* a
contract fallback for this, it triggers after 30 days, meaning the SN's
obligations get extended by an extra 15 days, which may be undeseriable
(for example, if someone has scheduled cancellation of a hosting plan).

This commit removes the revert for a repeated request so that if such a
situation does occur, the operator or a contributor can submit another
leave request call to attempt to propagate the event through to oxend
again.  Such duplicates can be helpful (as described above), but are
otherwise harmless: oxend simply ignores a leave request for an
already-unlocking node.

Such a subsequent request does *not* affect the leave request timestamp,
but merely reissues a leave request event.