exit/liquidate: Add specific error for non-existing BLS pubkey
If you call exit/liquidate with a BLS pubkey that doesn't exist in the
contract (for example, because it has just been removed) you currently
get a somewhat confusing `BLSPubkeyDoesNotMatch(0, pubkey)` error, where
the 0 is the result of a failed lookup.

This adds a more meaningful specific error for this case.