Atomic liquidation ratio update; remove >0 liquidator constraint
This replaces the three liquidator ratio updating functions with a
single function that updates all three values at once, so that this
value can be updated atomically.  Without this, you would need three
separate transactions, but in between those transactions being admitted
to the chain the ratio would be set to an undesired value.

This also adds a check that the liquidator+pool ratio will never exceed
1/3 of the recipient ratio, so that it will not exceed 1/4 of the total
stake because any higher and oxend is not guaranteed to be able to
deduct it from the operator.

This commit also removes the requirement that the liquidator ratio must
be strictly positive.  While it normally will be positive, from team
discussions we realized that there may be occasions to temporarily
disable the liquidation penalty, such as a blockchain bug leading to
significant numbers of deregistrations during which we want to disable
liquidation penalties without pausing the whole contract.